Export the NotificationsTable component
