+++ packages/atlas/src/components/NotificationsTable/index.ts
@@ -1,0 +1,1 @@
+export * from './NotificationsTable'
